requirements.txt
Flask>=3.0.0
Werkzeug>=3.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
gunicorn>=21.0.0